requirements.txt

whFlask==3.1.0
pillow==11.1.0
requests==2.32.3
gunicorn


